Don't retain connection used during gem initialisation
When the gem is initialised and a schema adapter is being used, the
adapter will initialise the schema_search_path which involves
touching the database, and thus checking out a connection for the
current thread. For threaded web servers, it is not expected that
the main thread has a database connection, so you typically set the
connection pool size equal to the number of web server threads.

This patch checks the connection back in to the pool after init so
that a web server thread can pick it up and use it.